Move tests from test_latex to test_to_latex
